requirements.txt
python-telegram-bot==20.3
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
